Shuffle controls can run single-threaded locally (no yeti)
setup_shuffle_model updated.
create_shuffle_configs updated.
exec_shuffle_model updated.
