+++ expt/shuffled_control_crfs.py
@@ -5,11 +5,10 @@
 import time
 import sys
 import os
-import stat
 import subprocess
 
 import crf_util
-from gridsearch_train_crfs import get_best_parameters
+import gridsearch_train_crfs
 import yeti_support
 
 import logging
@@ -51,6 +50,10 @@ def get_conditions_metadata(condition):
     metadata['shuffle_experiment_dir'] = os.path.join(
         params['source_directory'], "expt", metadata['shuffle_experiment'])
     params.update(metadata)
+    params["create_shuffles"] = create_shuffles
+    # There is no training prep by default, so we set no-op function as placeholder
+    params["shuffle_training_prep"] = lambda params: None
+    params["train_controls"] = exec_shuffle_model
 
     # Update settings for cluster specified, if any
     if params["cluster_architecture"] == "yeti":
@@ -60,64 +63,13 @@ def get_conditions_metadata(condition):
     return params
 
 
-def write_shuffling_yeti_script(params):
-    # write yeti script
-    filepath = os.path.join(params['shuffle_experiment'], "shuffle_yeti_config.sh")
-    logger.debug("writing file: {}".format(filepath))
-    with open(filepath, 'w') as f:
-        f.write("#!/bin/sh\n")
-        f.write("#shuffle_yeti_config.sh\n")
-        f.write("#PBS -N Create_shuffled_dataset_{}\n".format(params['shuffle_experiment']))
-        f.write("#PBS -W group_list={}\n".format(params['group_id']))
-        f.write("#PBS -l nodes={}:ppn={},walltime={},mem={}mb\n".format(
-            params['yeti_gen_sh_nodes'], params['yeti_gen_sh_ppn'],
-            params['yeti_gen_sh_walltime'], params['yeti_gen_sh_mem']))
-        if params['email_notification'] == "num_jobs":
-            # Always only 1 job, fully notify
-            f.write("#PBS -m abe\n")
-        else:
-            # Use email_notification setting verbatim
-            f.write("#PBS -m {}\n".format(params['email_notification']))
-        f.write("#PBS -M {}\n".format(params['email']))
-        f.write("#set output and error directories (SSCC example here)\n")
-        log_folder = "{}/".format(os.path.join(params['shuffle_experiment_dir'], "yeti_logs"))
-        os.makedirs(os.path.expanduser(log_folder), exist_ok=True)
-        f.write("#PBS -o localhost:{}\n".format(log_folder))
-        f.write("#PBS -e localhost:{}\n".format(log_folder))
-        f.write("#Command below is to execute Matlab code for Job Array (Example 4) " +
-                "so that each part writes own output\n")
-        f.write("matlab -nodesktop -nodisplay -r \"dbclear all;" +
-                " addpath('{}');".format(os.path.join(params['shuffle_experiment_dir'])) +
-                "gn_shuff_data('{}', '{}', {});".format(
-                    os.path.join(params['data_directory'], params['data_file']),
-                    params['shuffle_save_dir'],
-                    params['num_shuffle']
-                ) +
-                " exit\"\n")
-        f.write("#End of script\n")
-    f.closed
-    # Set executable permissions
-    os.chmod(filepath, stat.S_IXUSR | stat.S_IXGRP | stat.S_IXOTH | os.stat(filepath).st_mode)
-    logger.info("done writing {}".format(filepath))
-
-
-def write_shuffling_submit_script(experiment):
-    # write submit script
-    filepath = os.path.join(experiment, "shuffle_start_job.sh")
-    logger.debug("writing file: {}".format(filepath))
-    with open(filepath, 'w') as f:
-        f.write("qsub shuffle_yeti_config.sh\n")
-    f.closed
-    # Set executable permissions
-    os.chmod(filepath, stat.S_IXUSR | stat.S_IXGRP | stat.S_IXOTH | os.stat(filepath).st_mode)
-    logger.info("done writing {}".format(filepath))
-
-
-def write_shuffling_script(params):
-    write_shuffled_data_generating_script(params)
-    write_shuffling_yeti_script(params)
-    write_shuffling_submit_script(params['shuffle_experiment'])
-    logger.info("done writing {} yeti scripts\n".format(params['shuffle_experiment']))
+def create_shuffles(params):
+    matlab_cmd = "gn_shuff_data('{}', '{}', {});".format(
+        os.path.join(params['data_directory'], params['data_file']),
+        params['shuffle_save_dir'],
+        params['num_shuffle']
+    )
+    crf_util.run_matlab_command(matlab_cmd, add_path=params["source_directory"])
 
 
 def setup_shuffle_model(params):
@@ -126,24 +78,9 @@ def setup_shuffle_model(params):
     start_logfile(**params)
 
     os.makedirs(os.path.expanduser(params['shuffle_save_dir']), exist_ok=True)
+    # TODO: Either clear pre-existing shuffled datasets, or skip regenerating any already there
 
-    write_shuffling_script(params)
-
-    curr_dir = os.getcwd()
-    logger.debug("curr_dir = {}.".format(curr_dir))
-    os.chdir(params['shuffle_experiment'])
-    logger.debug("changed into dir: {}".format(os.getcwd()))
-
-    shell_command = ".{}shuffle_start_job.sh".format(os.sep)
-    logger.debug("About to run {}".format(shell_command))
-    process_results = subprocess.run(shell_command, shell=True)
-    if process_results.returncode:
-        logger.critical("\nAre you on the yeti cluster? Job submission failed.")
-        raise RuntimeError("Received non-zero return code: {}".format(process_results))
-    logger.info("Shuffled dataset creation job submitted.")
-
-    os.chdir(curr_dir)
-    logger.debug("changed back to dir: {}".format(os.getcwd()))
+    params["create_shuffles"](params)
 
 
 def create_shuffle_configs(params, best_params):
@@ -153,8 +90,14 @@ def create_shuffle_configs(params, best_params):
         f.write("    'datapath', '{}.mat', ...\n".format(
             os.path.join(params['shuffle_save_dir'], params['shuffle_save_name'])))
         f.write("    'experiment_name', '{}', ...\n".format(params['shuffle_experiment']))
-        f.write("    'email_for_notifications', '{}', ...\n".format(params['email']))
-        f.write("    'yeti_user', '{}', ...\n".format(params['username']))
+        try:
+            f.write("    'email_for_notifications', '{}', ...\n".format(params['email']))
+        except KeyError:
+            logger.debug("No notifications email setting provided. Skipping.")
+        try:
+            f.write("    'yeti_user', '{}', ...\n".format(params['username']))
+        except KeyError:
+            logger.debug("No yeti username provided. Skipping.")
         f.write("    'compute_true_logZ', false, ...\n")
         f.write("    'reweight_denominator', 'mean_degree', ...\n")
         # f.write("    's_lambda_splits', 1, ...\n")
@@ -184,91 +127,24 @@ def create_shuffle_configs(params, best_params):
     crf_util.run_matlab_command("write_shuffle_configs_for_{},".format(MODEL_TYPE),
                                 add_path=params['source_directory'])
 
-    create_controls_yeti_config_sh(params)
-    create_start_jobs_sh(params['shuffle_experiment'])
+    params["shuffle_training_prep"](params)
 
     os.chdir(curr_dir)
     logger.debug("changed back to dir: {}".format(os.getcwd()))
 
 
-def create_controls_yeti_config_sh(params):
-    # Expect to be in the experiment folder already when writing this
-    # TODO: Absolute path from source_directory
-    fname = "controls_yeti_config.sh"
-    with open(fname, 'w') as f:
-        f.write("#!/bin/sh\n")
-        f.write("#{}\n\n".format(fname))
-        f.write("#Torque script to run Matlab program\n")
-
-        f.write("\n#Torque directives\n")
-        f.write("#PBS -N Shuffled_CRFs_{}\n".format(params['shuffle_experiment']))
-        f.write("#PBS -W group_list={}\n".format(params['group_id']))
-        f.write("#PBS -l nodes={}:ppn={},walltime={},mem={}mb\n".format(
-            params['yeti_sh_ctrl_nodes'], params['yeti_sh_ctrl_ppn'],
-            params['yeti_sh_ctrl_walltime'], params['yeti_sh_ctrl_mem']))
-        num_jobs = int(params['num_shuffle'])
-        if params['email_notification'] == "num_jobs":
-            # Reduce email notifications for greater numbers of jobs
-            if num_jobs == 1:
-                f.write("#PBS -m abe\n")
-            elif num_jobs <= int(params['email_jobs_threshold']):
-                f.write("#PBS -m ae\n")
-            else:
-                f.write("#PBS -m af\n")
-        else:
-            # Use email_notification setting verbatim
-            f.write("#PBS -m {}\n".format(params['email_notification']))
-        f.write("#PBS -M {}\n".format(params['email']))
-        f.write("#PBS -t 1-{}\n".format(int(num_jobs)))
-
-        working_dir = os.path.expanduser(params['shuffle_experiment_dir'])
-        f.write("\n#set output and error directories (SSCC example here)\n")
-        f.write("#PBS -o localhost:{}/yeti_logs/\n".format(working_dir))
-        f.write("#PBS -e localhost:{}/yeti_logs/\n".format(working_dir))
-
-        f.write("\n#Command below is to execute Matlab code for Job Array (Example 4) so that " +
-                "each part writes own output\n")
-        f.write("cd {}\n".format(params['source_directory']))
-        f.write("./run.sh {0} $PBS_ARRAYID > expt/{0}/job_logs/matoutfile.$PBS_ARRAYID\n".format(
-            params['shuffle_experiment']))
-        f.write("#End of script\n")
-    f.closed
-
-    # make sure file is executable:
-    os.chmod(fname, stat.S_IXUSR | stat.S_IXGRP | stat.S_IXOTH | os.stat(fname).st_mode)
-    logger.info("Created " + fname + ".")
-
-
-def create_start_jobs_sh(experiment):
-    # Expect to be in the experiment folder already when writing this
-    # TODO: yeti specific
-    fname = "start_jobs.sh"
-    with open(fname, 'w') as f:
-        # Clear out any basic remainders from previous runs
-        f.write("rm -f ./results/result*.mat\n")
-        f.write("rm -f ./yeti_logs/*\n")
-        f.write("rm -f ./job_logs/*\n")
-        f.write("cd ../.. && qsub {}\n".format(
-            os.path.join("expt", experiment, "controls_yeti_config.sh")))
-    f.closed
-
-    # make sure file is executable
-    os.chmod(fname, stat.S_IXUSR | stat.S_IXGRP | stat.S_IXOTH | os.stat(fname).st_mode)
-    logger.info("Created " + os.path.join(experiment, fname) + ".")
-
-
-def exec_shuffle_model(shuffle_experiment, **kwargs):
+def exec_shuffle_model(source_directory, num_shuffle, **kwargs):
     curr_dir = os.getcwd()
     logger.debug("curr_dir = {}.".format(curr_dir))
-    os.chdir(shuffle_experiment)
+    os.chdir(source_directory)
     logger.debug("changed into dir: {}".format(os.getcwd()))
-
-    process_results = subprocess.run(".{}start_jobs.sh".format(os.sep), shell=True)
-    if process_results.returncode:
-        logger.critical("\nAre you on the yeti cluster? Job submission failed.")
-        raise RuntimeError("Received non-zero return code: {}".format(process_results))
-    logger.info("Shuffle jobs submitted.")
-
+    for cur_shuffle in range(1, num_shuffle + 1):
+        process_results = subprocess.run(".{}run.sh {}".format(os.sep, cur_shuffle), shell=True)
+        if process_results.returncode:
+            logger.critical("Training failed on shuffled control {}".format(cur_shuffle))
+            raise RuntimeError("Received non-zero return code: {}".format(process_results))
+        logger.info("Trained shuffled control model {} out of {}.".format(cur_shuffle,
+                                                                          num_shuffle))
     os.chdir(curr_dir)
     logger.debug("changed back to dir: {}".format(os.getcwd()))
 
@@ -311,13 +187,13 @@ def main(condition):
     # Create bare-bones shuffle folder and create shuffled datasets
     setup_shuffle_model(params)
     # Get best params
-    best_params = get_best_parameters(params['experiment'])
-    logger.info("Parameters for all conditions collected.\n")
+    best_params = gridsearch_train_crfs.get_best_parameters(params['experiment'])
+    logger.info("Parameters for {} collected.\n".format(condition))
     # create shuffle configs with best params (write and run write_configs_for_loopy.m)
     create_shuffle_configs(params, best_params)
     # Wait for all shuffled datasets to be created and run shuffle/start_jobs.sh
     params['to_test'] = simple_test_shuffle_datasets
-    params['to_run'] = exec_shuffle_model
+    params['to_run'] = params["train_controls"]
     crf_util.wait_and_run({condition: params})
     # Wait for shuffle CRFs to be done and run merge and save_shuffle
     params['to_test'] = test_shuffle_CRFs
